Local commit

+++ mozmill/mozmill/extension/nativeEvents/native_events.c
@@ -1,6 +1,6 @@
 #include "native_events.h"
 
-guint sendClick(gdouble x, gdouble y, guint button) {
+guint sendClick(gint x, gint y, guint button) {
     return click(x, y, button);
 }
